+++ src/components/Legend/index.js
@@ -5,7 +5,7 @@ import {
   Dino,
   LegendGroup,
 } from './styled';
-import { dinos } from '../data/dinos.js';
+import dinos from '../../data/dinos.js';
 
 const Legend = ({ color, reversed, legend }) => {
   const legendData = Object.entries(legend);
